Add Workspaces folder inside pages Folder and embed subworkspace pages + rename pageTeam and PageDomainChild to pageWorkspace and pageWorkspaceChild
